feat: update layout and chat components for Course Architect with enhanced UI and new ChatPanel
